requirements.txt
GitPython
hjson
tabulate
pyyaml
mako
sphinx
sphinx-rtd-theme
recommonmark
sphinxcontrib-svg2pdfconverter
